fix(jangar): split agentrun ingestion proxy
